refactor: outsource product detail fetch logic in a custom hook
